Add README.md
Chapter 1: Multi-Currency Money
-------------------------------

We are tasked with creating software to model investments in different
currencies.  For example we need to be able to produce reports like
this:

    Instrument | Shares | Price   | Total
    -----------+--------+---------+----------
    IBM        | 1000   | 25 USD  | 25000 USD
    Novantis   | 400    | 150 CHF | 60000 CHF
    -----------+--------+---------+----------
                        | Total   | 65000 USD

The overall total is calculated using some exchange rate:

    From | To  | Rate
    -----+-----+-----
    CHF  | USD | 1.5

What tests can we come up with to drive the development of code which is
able to produce such a report?