Add first multiplication test
We start working with the easiest task on our to-do list -- simple
multiplication of single currency. We mark the current item on our list
to keep us focused.

The test at this point will not even compile, as we haven't added the
`./money.js` module or the `Dollar` class. We'll work incrementally, and
handle each error as it comes, taking our cues from the test runner.