Add first implementation of `Dollar.times()`
Our task here was to get the test to pass as quickly as possible,
writing any code at all that will achieve that, no matter how stupid or
dirty.

We fixed the following steps to get the test to compile:

1. Added the `./money.js` module
2. Exported the `Dollar` class from `./money.js`
3. Declared the `Dollar.times()` method

To get the test to pass, we hardcoded `Dollar.amount = 10`.