Refactoring step 1
In the refactoring step we aim to reduce duplication in the code.

Here the duplication was the value `10` (or `5 * 2`) between the code
under test and the test itself `Dollar(5).times(2)`.

This duplication of the value `10` is a symptom of coupling between the
code and the test. If we were to write another multiplication test, say
something like `3 USD * 3 = 9 USD`, then we would have to change our
hard-coded value of `10`, and then go on to write code which would make
*both* tests pass.

If we refactor *now* to reduce coupling (indicated by duplication), then
we set ourselves up to be able to write a new test and make it pass with
the smallest and simplest possible change.

**We are constantly making our software easier to change**

The refactoring stage can be done in very small steps, for example here
we just make explicit the duplication between the test and the code, and
move some logic into the `times` method.

In the following few refactoring steps we will aim to reduce this
duplication completely.

After each change we run the tests again to make sure we haven't broken
anything.

Once you are feeling confident, you can take larger steps (using a
higher "gear" as Beck puts it). But it's good to practice taking the
smallest possible steps to prepare for when you are dealing with
something really difficult.