Refactoring step 2
Here we move initialisation of `amount` into the constructor. Our
eventual aim is to totally remove `5` and `2` from our code, and so to
eliminate the duplication with the test code.

We are taking another tiny step here, noticing that `5` was the number
passed to the constructor, and `2` the argument passed to `times`, so we
are setting ourselves up to replace the hard-coded values with
arguments in the next refactoring.