Refactoring step 3
Here we have completed our goal of eliminating coupling between the test
and our code. All references to `5` and `2` have been removed.

This will make it much easier for us to make changes in the future.
Refactor to remove duplication

To recap on our process so far, we have:

1. Chosen the easiest task from our to-do list
2. Written a simple test to express how we would like our code to behave
3. Made the test run, and seen it fail
4. Made a little dirty change in the code to make the test pass
5. Refactored the code to remove duplication

This is the **TDD cycle**, and should only take a couple of minutes to
complete.

**Note** that in this case, the duplication was between the code and the
test, but it's often the case that we find duplication between two
pieces of code, which is easier to spot.

In any case, by diligently removing duplication in every cycle, we
maximise our opportunity to get our next test running with one single
small change.