Add test for new `Dollar.times()` implementation
Here we start to tackle the issue of side effects from multiplication.

We modify our existing test to reflect how we would like to be able to
use the `Dollar` class. It's fine to change our tests as we go, we are
unlikely to choose the perfect interface for our code on the first try,
just as we're unlikely to immediately choose the perfect implementation.