Implement `Dollar.times` without any side effects
Last time we "committed some sins" as Beck puts it, to make our test
pass. We did this to get to a passing state as quickly as possible,
which then gives us room to explore a cleaner implementation from a
place of safety.

This time the implementation is obvious, and so we can just write it in.

We have now seen two of the three strategies Beck outlines for quickly
getting tests to pass:

    1. "Fake it -- Return a constant and gradually replace constants
       with variables until you have the real code."

    2. "Use Obvious Implementation -- Type in the real implementation."

Which one you choose depends on your level of confidence with the code
you're writing.

The third strategy is `triangulation`, which we will deal with soon.

To review the steps we just took:

    1. We took a feeling about the design of our code (our to-do) and
       translated it into a failing test.

    2. We made the test work by typing in an obvious implementation.

The first step -- translating our feelings about the design of our code
into to-do items, and then into failing tests -- is a recurring theme of
TDD. It allows you to reflect on the design of your code *first*, and
only afterwards worrying about the best way to implement the correct
behaviour.