Update TODO.md
Since we now return a new `Dollar` instance every time we change a
dollar's amount, we have now eliminated the risk of *aliasing*, that is
inadvertently modifying one object while modifying another.

This is a key component of the *Value Object* pattern -- the value of an
object is fixed for its lifetime and guaranteed not to change.

An implication of this is that our value objects must implement
`equals()`, since two `Dollar`s of equal value should be interchangable.

Another implication of the Value Object pattern is that objects can be
used as the keys to a hash table (in Java: `HashMap`, in Python:
`dictionary`, in JavaScript: `Map`).

In the book, Beck adds an implementation of `Dollar.hashCode()` to his
to-do list, in order to allow this functionality. In JavaScript, we can
use any value (objects and primitives) as a key in a `Map` without
having to define this behaviour, so we don't bother with it.

Now, on our to-do list, we cross off our previous task, and start work
on `Dollar.equals()`.