Add `Dollar.equals` method.
We add a straightforward test for equality between `Dollar`s of the same
amount, and implement a fake implementation to get the test to pass.

We are being deliberately obtuse in this case in order to introduce the
third, and most conservative strategy of TDD: *triangulation*.

Triangulation in this context means to defer generalisation of the code
until we have two examples or more, analogous to triangulation of radio
signals, which requires information from two or more receivers to
pinpoint the location of a signal.

Generally, triangulation is intended for when we don't have a good idea
of how to generalise our code. But, for the purposes of demonstration,
in this case we will ignore the duplication of the value `true` between
our test and code, and refrain from generalising our implementation
until we are forced to change it by a new test.