Generalise `Dollar.equals()` implementation.
By adding out new test we use triangulation to force ourselves to
generalise `Dollar.equals()`.

Incidentally, we could have also used triangulation to drive our
implementation if `Dollar.times()` as well -- If we had `$5 x 2 = $10`
and `$5 x 3 = $15` then we wouldn't have been able to return a constant.

In general, if you can see how to eliminate duplication between the code
and the test, and implement a general solution, then you should do it.

However, as previously mentioned, this usage of triangulation is quite
contrived, and in general its use is reserved for when implementation or
design ideas are not forthcoming. By giving yourself more concrete
examples of the desired behaviour of your system, you can often better
conceive of a solution.