Make `Dollar.amount` private.
Now we have an implementation for `Dollar.equals()` we can rewrite our
test for mulitiplication to remove all references to `Dollar.amount`, as
well as inlining the temporary `product` variable.

Now our test reads a little clearer and more concisely, and since the
only references to `Dollar.amount` are from within the `Dollar` class,
we can make `Dollar.amount` private.

It's important to note here that we introduce a dependency on
`Dollar.equals()` to our test for `Dollar.times()`. This means that any
errors in `Dollar.equals()` would break our test for multiplication as
well.

This is a risk that we should be aware of, and actively manage in TDD.
We can't achieve perfection, only confidence in our code. From time to
time a defect will creep in and bite us in the backside. When this
happens we need to write a test to cover the behaviour we should have
implemented to prevent the defect and move on.

In review, for this change we:

    1. Used newly-implemented functionality to refactor a test.
    2. Noticed the dependency of one test upon another.
    3. Decided to proceed in spite of this risk.
    4. Used the new functionality to reduce coupling between the test
       and the code (making an attribute private).