Add Franc class.
We want to work towards completing our first to-do item:

    `$5 + 10 CHF = $10 if rate is 2:1`

This seems still a little complicated, but we can imagine that we need a
class `Franc`, similar to `Dollar`, which will bring us one step closer
to being able to satisfy this requirement.

And so we update our to-do list with a requirement for the `Franc`
class, analogous to that of the `Dollar` class:

    `5 CHF * 2 = 10 CHF`

In order to satisfy this, we can happily copy-paste our `Dollar` code,
and change `Dollar` to `Franc`. Easy, right?

We immediately cross off the new to-do item as it was trivial to
implement.

This blatant duplication can and will be fixed, but for now we must
respect the TDD cycle and accept the quick-and-dirty solution. We will
work hard in the next commits to reduce this duplication and generalise
the behaviour of these two currencies, starting with the `equals()`
method. We update our to-do list accordingly.

To review, in this commit, we:

    1. Couldn't tackle the big test, so wrote a smaller test to get us a
       little closer.
    2. Wrote and satisfied the test by shameless duplication.
    3. Dedicated ourselves to removing this duplication in the coming
       commits.