Add `Money.equals()`.
Thanks to our preparatory work in the previous commits, we can make this
change very easily. We first copy the `equals()` method up into `Money`,
rename the parameter from `dollar` to `money` for clarity, and then
delete `Dollar.equals()`.

We can now move on to doing the same for `Franc.equals()`.