Add `Money.amount` property.
Our end goal is to move the `equals()` method up into `Money`, but first
we need to move `amount`, since `equals()` depends upon it.

Here we take the first step and move the declaration of the `amount`
property from `Dollar` to `Money`, making it public instead of private
to avoid breaking `Dollar`.