Add forgotten tests for `Franc.equals()`.
We start thinking about our next task, moving `Franc.equals()` up into
`Money`, but oh no! We suddenly realise that we never wrote any tests
for `Franc` equality! We must have forgotten when copy-pasting.

It is often the case when implementing TDD (especially when workng
with legacy code) that you discover you don't have adequate tests.

When this happens, just write the tests that you wish you had. Then you
have a benchmark for refactoring. See also Dave Farley on *acceptance
testing*.

In this case, we can just copy-paste again to add the tests for
`Franc.equals()`, and move on.