Make `Money.equals()` handle comparing different subclasses equal.
It occurred to us when working on generalising `equals` that we haven't
decided what happens if we compare `Franc`s to `Dollar`s. We dutifully
write a test for this and it fails, 5 `Franc`s equals 5 `Dollar`s.

To make the test pass we add a crude check that the class name of both
objects is the same. Testing for the type of an object in this way is
usually considered bad practice. We suspect that at some point we
will need to replace this check with a check for equal currencies, but
so far we haven't introduced a concept of currency, and so we will leave
it for now.

To review what we did in this commit:

    1. We noticed something about our code which bothered us and wrote a
       test for it.
    2. Made the test run in a reasonable but slightly smelly way
       (testing class names).
    3. Decided to put off adding to our design before we have more
       motivation to do so.