Add `Money.dollar()` factory method.
We want to eliminate the `Dollar` and `Franc` subclasses, as they both
only have a single method left, `times()`, and both subclasses'
implementations of that method are very similar.

However it's not immediately clear how we can go about this, so we
choose as a small step in that direction to remove all direct references
to the subclasses in our code.

We can start by introducing a `Money.dollar()` factory method which will
return a new `Dollar` instance.

We replace all references to `Dollar` in our tests with `Money.dollar`,
and write the obvious implementation to satisfy the tests.

Note that the factory method `Money.dollar` is *static* as it is called
from the class itself and not an instance of that class.

Now when the time comes to eliminate the `Dollar` subclass we won't have
to change any client code, since we have decoupled our tests from the
existence of `Dollar`.