Add `Money.franc()` factory method.
We repeat the previous steps but for `Franc` -- We create a factory
method on `Money` and eliminate all references to `Franc` from our
test code.

In the process we notice that the test "Franc multiplication" is
actually testing the same code as "Dollar multiplication", since they
both delegate `equals` to the `Money` class.

However, we refrain from deleting this test as we are still not entirely
convinced that it's reduntant yet. Our criteria for deleting a test
should be that it does not affect our confidence in our code if we
remove it. If we're still suspicious, it's better to leave it in.

We don't cross off this task from our to-do list as we didn't complete
it yet, but we made some small progress at least.

In review, this commmit we:

    1. Decoupled the existence of our subclasses from our test code.
    2. Noted that when the subclasses disappear, some test code will
       become redundant, but did not delete anything yet.