Add `currency` property to `Dollar` and `Franc`.
We want to work towards eliminating the `Dollar` and `Franc` subclasses,
so what on our to-do list would help us achieve that goal? We could
replace the subclasses with a simple string property on the `Money`
class which represents either a dollar ("USD") or a franc ("CHF").

We dutifully write a test for our desired behaviour, and make it pass
very stupidly by adding static properties to `Dollar` and `Franc` with
the desired values.