Refactoring 2.
We want to move the currency strings "USD" and "CHF" out of the
subclasses and into the superclass. We could try setting these strings
in the `Money.dollar` and `Money.franc` factory methods, which would
require a change in the constructor signature of the subclasses.

Our first step is to make "USD" an instance attribute instead of a class
attribute, and initialise it in the `Dollar` constructor.