Refactoring 3.
Before we change the `Dollar` constructor's signature, we notice that
`Dollar.times()` calls this constructor, and so we need to change this
call first to prevent calling the constructor with the incorrect number
of arguments.

JavaScript, in its infinite wisdom, does not consider passing the
incorrect number of parameters to a function to be an error, however we
can consider it bad practice and avoid it!