Refactoring 6.
Now that the subclasses' constructors are identical we can generalise
and move them up into the superclass. We can also move the `currency()`
getter as well now. We didn't do this before because `_currency` wasn't
declared in `Money` yet.

Now all that's left of the subclasses are rather similar-looking
implementations of `times()` -- we update our to-do list and prepare to
generalise `times()`.

To review the changes we made in these last commits:

    1. We were stuck when trying to remove duplication between `Dollar`
       and `Franc`, so we worked on an easier to-do item which we
       thought might help us.
    2. We reconciled differing constructor implementations in `Dollar`
       and `Franc` by moving the differences to `Money`.
    3. Repeated our slow and careful changes to `Dollar` in one big
       change to `Franc` once we were feeling more confident.
    4. Pushed up the identical constructors to the parent class and got
       one step closer to eliminating our subclasses.